requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pymupdf==1.27.2
chromadb==0.5.15
google-generativeai==0.8.3
python-dotenv==1.0.1
langchain-text-splitters==0.3.0
jinja2==3.1.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic[email]==2.12.4
python-docx==1.1.2